fix(bayn): preserve production dependency layering
